requirements.txt
aiohttp==3.9.1
atproto==0.0.20
tweepy==4.14.0
facebook-sdk==3.1.0
pillow==10.1.0
praw==7.7.1
python-linkedin-v2==0.9.4
instabot==0.117.0
schedule==1.2.1
folium==0.15.0
cartopy==0.22.0
matplotlib==3.8.2
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
feedparser==6.0.10
sqlalchemy==2.0.23
colorama==0.4.6
tqdm==4.66.1